Studio ad-don settings/columns explorer: dialog.Show -> dialog.ShowModal.
When closing columns explorer settings dialogs, focus moves to OK button in SPL add-on settings. This is because dialog.Show is called instead of dialog.ShowModal (columns explorer dialogs shoud be modal dialogs). Changing dialog.Show to dialog.ShowModal resolves the focus issue and makes the design of the dialogs clearer. Also, as part of this work, self.Disable call on SPL add-on settings are removed for columns explorer panel.