SPL common services: create a copy of SPL base services module inside splcommon package.
The SPL common services (splcommon) is a package of base modules used across both the app modules set and the global plugin and support modules for them. The aim is to house common code across the add-on in hopes of reducing code duplication. This also means the common services package is responsible for overall functioning of the add-on. To acknowledge that Studio ap module is at the top of the app module set/global plugin hierarchy for this add-on, splcommon package is placed inside appModules folder with initial code coming from the Studio app module components used across the add-on.
Note: this is a continuation of recent code refactoring and restructuring efforts including Python 3.11 upgrade, removing outdated code comments, and Studio 6.x requirement.