Metadata streaming: let the splmisc module handle streaming connecoitn and announcement at startup.
NVDA can be told to connect and announce metadata streaming status when Studio (more specifically, the app module) starts. However, this routine is filled with calls to private functions inside pslmisc module. Therefore, move this routine to splmisc module so private functions can remain private (and perhaps combine this routine with others involving streaming connection and reporting in the future). As part of this change, Studio init event is passed into the splmisc version of the startup handler so the event can be set to 'wiat' status.