Studio app module/track finder dispatch: remove unnecessary find text filtering procedure.
The idea of filtering track finder text was to remove duplicate or invalid entries when find next or previous is called. However, this is unnecessary and did not work properly because find text attribute is a list, not a string. Therefore, remove filtering procedure (it may return in the future when track find text becomes a string once again (retiring search history/list).