Local Studio: search for Studio window handle if the API is truly required.
Local and Remote Studio differences: for now, Remote Studio does not require Studio API (there might be its own API). Therefore, only look for Studio window handle if this is local Studio (Studio API required flag is True). This is applicable for app module constructor and terminate methods (applies to scripts but that is next commit).