Local and Remote Studio: weather and temperature is unavailable in Remote Studio.
According to Remote Studio documentation and the user interface, weather and temperature is unavailable in Remote Studio. The newly introduced local Studio only decorator cannot be used in the weather script because an input gesture can be added to this scipt, and this is done with another decorator. Thus, adding a second (nested) decorator will not work - for instance, input hep message for weather and temperature command will not be announced. Therefore, add a script specific check and return if Remote Studio is in use (and announce that this script is unavailable in Remote Studio).