Reote Studio/name change: create a private method to check if track time is below the alarm threshold.
Remote Studio might not have an API. As long as this is the case, screen content must be relied upon when handling end of track/song intro alarms. Therefore, a private method will be used to make sure rack time display is below the alarm threshold. This method converts time display to seconds and compares it with alarm value, returning True if the time display is at or below the end of track/song ramp threshold.