Studio/time range finder: forcefully convert (shorten) tracks more than an hour long into hour long tracks.
If the track is more than an hour long, it will be hsown as 'hh:mm:ss'. This causes Python to say that these tracks are shorter than the maximum value possible (59:59). Therefore, forcefully convert (shorten) these tracks to hour long tracks so that minimum <= duration ,+ maximum can continue to work for string comparisons.