Local and Remote Studio/playlist viewer item: 'Time Scheduled' -> 'Time' when building custom column header and content combo in local Studio 6.20.
Change 'Time Scheduled' to 'Time' when building custom column header/content combo in local Studio 6.20. In order to do so and to guarantee backward compatibility, included columns and column order must be copied because 'Time Scheduled' will be changed to 'Time', and just modifying SPL config database can cause errors when a broadcaster decides to downgrade from local Studio 6.20 to 6.11.