Studio: prevent many microphone monitor threads from running when enabling microphone.
Unlikely but possible: both local Studio and Remote Studio might be running on the same computer. When toggling microphone, there will be a lag from one Studio variant to another. When microphone is then switched off and on again, the mic alarm thread will not realize that microphone was off and back on. This will cause many mic reminder threads to run. Therefore, when microphone is enabled, return early if mic alarm thread is already running.