SPL utilities: return SPL Controller scope and actual active Studio component via a dedicated method. Re #169.
Define a dedicated method to return SPL Controller scope option and active Studio component. The method first opens SPL config database by checking for existence of any Studio top-level window (TStudioForm), and if not, it will return no scope and no active component (None, None). If the config was opened successfully (a Studio window is locaed), it will compare the scope with the list of active Studio components:
* If scope is None (both local and Remote Studio), the first active component will be returned as active component.
* If a specific scope is defined (splstudio or remotestudio), the scope setting will be returned if the component of the same name is runing 9scope is active), otherwise None will be returned.
Returning None for active Studio component ensures that SPL Controller layer and user assignable scripts do not proceed even if SPL Controller scope is defined (or not).