SPL utilities: define a decorator to limit some SPL COntroller layer commands to local Studio. Re #169.
Similar to local Studio app module: define a decorator that will play an error tone (as if SPL COntroller command is unassigned to a key) if active Studio component is Remote Studio (either Remote Studio is the only active app while scope setting is none (both), or a broadcaster might be running both local and Remote Studio but chose to limit SPL Controller to Remote Studio). This will be applied to commands unavailable from Remote Studio itself (next commit).