SPL track item (base class)/columns explorer: ask concrete track item classes if on-screen column headers have changed.
Reported by a broadcaster: if 'Time Scheduled' is set as a columns explorer slot in local Studio 6.20, NVDA will say 'Time Scheduled not found'. This is because in local Studio 6.20. 'Time Scheduled' is 'Time'. Resolving this requires telling SPL track item classes that the new on-screen column header should be used, but columns explorer is defined inthe base class. Therefore, let the base SPL track item class ask ocncrete classes if column header for the columns explorer slot should be updated (this will be implemented for playlsit viewer class in local and Remote Studio).