Local and Remote Studio: remove the need to check for local Studio only flag when announcing current track title.
Remote Studio uses identical API to that of local Studio when reporting playback and other status information (for the status bar). Therefore, there is no need to limit current track title reporting to local Studio as it can be used in Remote Studio, particularly when announcing that no track is playing.