Local Studio: define screen object traversal routes when error log button is showing.
A button is shown in local Studio when an error occurs - the button will open a general Studio log. This button is located in the same place as other Studio UI controls, and thus displaces some items by 1 (one) in object navigation order. Therefore, define traversal routes for this case (6viewerrorlog) to let NVDA announce information when the view log button is showing.