Remote Studio: check simple grandparent, not window style when looking for playlist viewer items.
Over the years, different window styles were used for playlist viewer items, and this is more so for local Studio. In the initial implementatoi, Remote Studio also checked window style. However, it might be possible that there are other list views with the same window style that are not supposed to be playlist viewer items. One sure way to check for the 'real' playlist viewer item is checking foregrond window class name (TStudioForm), and for Remote Studio, this can be obtained form simple grandparent object or the list item. Testing shows the grandparent approach works, so use it instead of window style checks when looking for playlist viewer items.