Local Studio: check simple grandparent window class name when looking for playlist viewer items.
Just like Remote Studio, instead of relying on window style, check simple grandparent object window class name. This also works for local Studoi because the simple grandparent of the playlist viewer item is Studio foregrond window (TStudioForm). This removes the need to check window style when new Studio versions are released but should be checked in case object hierarchy changes in the future.