Local Studio/insert tracks:lirary scan progress: handle a case when progress text is empty and library scan is finished according to Studoi API.
Sometimes, library scan progress text can become empty. If that happens, progress will not be nanounced if scan is indeed happening. However, there is a case when progress text is empty and scan is indeed complete. Therefore, catch the second case and terminate library scan progress announcer. This is effectively 'match' text check with library scan flag on.