Studio/SPL Assistant layer entry: rework gesture assignments for layer commands via 3-tuple lookup.
Define a named tuple of (script, NVDA layout, JAWS layout) and use it to parse 3-tuples defining SPL Assistant layer commands. Note that not all layer commands will be associated with keyboard gestures ('None' is recorded for layer commands without gestures defined), so only bind gestures if there is a command associated with the script. While sacrificin readabiilty somewhat, this makes SPL Assistant layer command assignments (end of the module) easier to read in the long-term.